server-func: separate browser and node esm exports
