+++ packages/start/package.json
@@ -29,6 +29,36 @@
     "error-boundary",
     "api"
   ],
+  "exports": {
+    ".": {
+      "types": "./plugin.d.ts",
+      "import": "./plugin.js"
+    },
+    "./addons/*": "./addons/*",
+    "./api": "./api/index.ts",
+    "./api/*": "./api/*",
+    "./entry-client": "./entry-client/index.tsx",
+    "./entry-client/*": "./entry-client/*",
+    "./entry-server": "./entry-server/index.ts",
+    "./entry-server/*": "./entry-server/*",
+    "./error-boundary": "./error-boundary/index.tsx",
+    "./error-boundary/*": "./error-boundary/*",
+    "./meta": "./meta/index.tsx",
+    "./meta/*": "./meta/*",
+    "./root": "./root/index.tsx",
+    "./root/*": "./root/*",
+    "./router": "./router/index.tsx",
+    "./router/*": "./router/*",
+    "./runtime/*": "./runtime/*",
+    "./server": {
+      "browser": "./server/browser.ts",
+      "node": "./server/server.ts",
+      "default": "./server/server.ts"
+    },
+    "./server/*": "./server/*",
+    "./session": "./session/index.ts",
+    "./session/*": "./session/*"
+  },
   "dependencies": {
     "@babel/core": "^7.16.7",
     "@babel/plugin-syntax-jsx": "^7.16.7",
