+++ packages/start/server/browser.ts
@@ -1,0 +1,5 @@
+import { server } from "./server-functions/browser";
+
+export default server;
+
+export * from "./shared";
