+++ packages/start/server/index.ts
@@ -1,10 +1,5 @@
-import { server } from "./serverFunction";
-
+// This file only exists to make the TypeScript tsconfig setting `"moduleResolution": "node"` work.
+// See ./server.ts for the server entrypoint and ./browser.ts for the browser entrypoint
+import server from "./server";
 export default server;
-
-export * from "./responses";
-export * from "./StartContext";
-
-export { StatusCode } from "./StatusCode";
-export { HttpHeader } from "./HttpHeader";
-export { createServerResource, createServerAction } from "./resource";
+export * from "./server";
