+++ packages/start/server/server-functions/babel.js
@@ -2,7 +2,7 @@
 // https://github.com/vercel/next.js/blob/canary/packages/next/build/babel/plugins/next-ssg-transform.ts
 // This is adapted to work with any server() calls and transpile it into multiple api function for a file.
 
-import { INLINE_SERVER_ROUTE_PREFIX } from "./constants.js";
+import { INLINE_SERVER_ROUTE_PREFIX } from "../constants.js";
 import nodePath from "path";
 import crypto from "crypto";
