+++ packages/start/server/server.ts
@@ -1,0 +1,5 @@
+import { server } from "./server-functions/server";
+
+export default server;
+
+export * from "./shared";
